fix landing
