+++ frontend/src/components/TestLanding.js
@@ -1,6 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
-import Form from "./Form";
+import ContactForm from "./ContactForm";
+import Login from "./Login";
+import Signup from "./Signup";
 
 import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
@@ -21,6 +23,12 @@ export default function App() {
             <Nav.Link>
               <Link to="/form">Contact us</Link>
             </Nav.Link>
+            <Nav.Link>
+              <Link to="/login">Log in</Link>
+            </Nav.Link>
+            <Nav.Link>
+              <Link to="/signup">Sign Up</Link>
+            </Nav.Link>
           </Nav>
         </Navbar>
 
@@ -33,6 +41,12 @@ export default function App() {
           </Route>
           <Route path="/">
             <Home />
+          </Route>
+          <Route path="/login">
+            <LoginPage />
+          </Route>
+          <Route path="/signup">
+            <SignupPage />
           </Route>
         </Switch>
       </div>
@@ -49,5 +63,11 @@ function About() {
 }
 
 function Contact() {
-  return <Form />;
+  return <ContactForm />;
+}
+function LoginPage() {
+  return <Login />;
+}
+function SignupPage() {
+  return <Signup />;
 }
